+++ src/main/java/kitchenpos/tablegroup/domain/TableGroupRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.table.domain;
+package kitchenpos.tablegroup.domain;
 
 import java.util.List;
 import java.util.Optional;
