+++ src/main/java/kitchenpos/tablegroup/dto/TableGroupCreateRequest.java
@@ -1,9 +1,10 @@
-package kitchenpos.table.dto;
+package kitchenpos.tablegroup.dto;
 
 import java.util.List;
 import java.util.stream.Collectors;
 import kitchenpos.table.domain.OrderTable;
-import kitchenpos.table.domain.TableGroup;
+import kitchenpos.table.dto.OrderTableRequest;
+import kitchenpos.tablegroup.domain.TableGroup;
 
 public class TableGroupCreateRequest {
