+++ src/test/java/kitchenpos/support/ServiceTest.java
@@ -10,9 +10,9 @@ import kitchenpos.product.domain.ProductRepository;
 import kitchenpos.product.domain.ProductService;
 import kitchenpos.support.repository.OrderMenuRepository;
 import kitchenpos.table.domain.OrderTableRepository;
-import kitchenpos.table.domain.TableGroupRepository;
-import kitchenpos.table.domain.TableGroupService;
 import kitchenpos.table.domain.TableService;
+import kitchenpos.tablegroup.domain.TableGroupRepository;
+import kitchenpos.tablegroup.domain.TableGroupService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.jdbc.Sql;
