+++ src/test/java/kitchenpos/table/TableServiceTest.java
@@ -23,9 +23,9 @@ import kitchenpos.support.fixtures.OrderLineItemFixtures;
 import kitchenpos.support.fixtures.OrderTableFixtures;
 import kitchenpos.support.fixtures.TableGroupFixtures;
 import kitchenpos.table.domain.OrderTable;
-import kitchenpos.table.domain.TableGroup;
 import kitchenpos.table.dto.OrderTableCreateRequest;
 import kitchenpos.table.dto.OrderTableResponse;
+import kitchenpos.tablegroup.domain.TableGroup;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
